+++ examples/custom-input-default-config/custom-input-default-config.vue
@@ -1,0 +1,14 @@
+<template>
+  <FormKit type="hello" />
+</template>
+
+<script type="formkit-config">
+{
+  inputs: {
+    hello: {
+      type: 'input',
+      schema: ['Hello world']
+    }
+  }
+}
+</script>