+++ examples/custom-input/custom-input.vue
@@ -1,0 +1,10 @@
+<template>
+  <FormKit :type="helloWorld" />
+</template>
+
+<script setup>
+const helloWorld = {
+  type: 'input',
+  schema: ['Hello world'],
+}
+</script>